requirements.txt
fastapi[standard]
langchain==0.3.9
langchain_openai==0.2.10
langchain_community==0.3.8
graphrag_sdk[all]==0.3.3
falkordb==1.0.10
python-dotenv==1.0.1
